Some mobile optimizations.
The main idea is to have some fallbacks for mobile devices, especially
android default browser.

Our main layer callded "terrain01" is too much bigger for the crappy
android browser, so we'll generate a new simple layer to render when the
game starts on this device.

I've added some pre-rendering for background elements, trying to improve
the performance. The level loading takes some seconds but the
performance seems to be better.